Move mock API classes to logic.api package
